requirements.txt
pandas>=2.0
pyarrow>=15.0
datasets>=2.18
transformers>=4.40
ftfy>=6.2
pyyaml>=6.0
matplotlib>=3.8
seaborn>=0.13
jupyter>=1.0
pytest>=8.0
tqdm>=4.66
